requirements.txt
gradio == 5.49.1
huggingface_hub == 1.1.2
tavily-python == 0.7.12
datasets == 4.4.1